dusd-dashboard/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
